Additional error checks
A couple of years ago, I used this lib to run the conformance tests at http://u2fconformance.fidoalliance.org. To comply with those tests, I had to add a few error checks to the U2F lib so that it doesn't return false positives. If I remember correctly, I took these error checks from the Java implementation and ported them over to PHP. I did this in June 2015, so I might be mis-remembering the source from which I took these changes. I just came back to this repo to check if there were any bug fixes and cleanups, and realized I had never posted this upstream, so here it is.