Support single files for attestation
Wrapping the `$dir` condition with `is_dir`, otherwise it throws a PHP warning if a string that isn't a directory is passed.
Adds an elseif to check if the path is a valid file, if so add it to the files list.

This is the minimal possible changeset for this, I can add tests if needed. I think `$attestDir` can be renamed to `$attestPath` to better match what it does, if this is a change you'd be open to merge. Please let me know if I should make those additional changes.